requirements.txt
kagglehub[pandas-datasets]
python-dotenv
openai
matplotlib
numpy
pandas
datasets
